KP-6210 Follow operator module division in tests
Commit c4cf8d34562d62d14 split the file downloading operators from
`custom_operators.py` into a module of their own. Now the tests are
similarly split.
